+++ Changes.md
@@ -3,11 +3,14 @@
 
 ### Changes
 
+- pass smtp.ini config from Server into connections & transactions
+
 ### New features
 
 - add ability to disable SMTPUTF8 advertisement #2866
 
 ### Fixes
+
 
 ## 2.8.26 - 2020-11-18